+++ src/main/java/dao/ServicerDao.java
@@ -93,8 +93,12 @@ public class ServicerDao implements IDao<Servicer>{
     }
 
     @Override
-    public Servicer delete(Integer id) {
-        Servicer servicer = getById(id).orElse(null);
+    public Optional<Servicer> delete(Integer id) {
+        Optional<Servicer> servicer = getById(id);
+        if(servicer.isEmpty()){
+            return Optional.empty();
+        }
+
         String query = "DELETE FROM User WHERE id = ?";
 
         try(PreparedStatement statement = conn.prepareStatement(query)) {
